requirements.txt
streamlit 
pandas 
numpy 
yfinance 
plotly.express 
alpha_vantage
stocknews  